requirements.txt
python-telegram-bot>=21.6
sqlalchemy>=2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.9
